utils for building package

+++ packages/extension/scripts/bun/dev.ts
@@ -1,16 +1,16 @@
 process.env.NODE_ENV = 'development'
 
-import chalk from 'chalk'
 import { watch } from 'fs'
 import fs from 'node:fs/promises'
 import { head, last, isString } from 'lodash'
 import yargs from 'yargs'
 import { execa } from 'execa'
 import { hideBin } from 'yargs/helpers'
-import stripJsonComments from 'strip-json-comments'
 import 'colors'
 
-import { get_config } from './config'
+import { build } from './utils'
+
+const OUT_DIR = 'out'
 
 const argv = yargs(hideBin(process.argv)).parseSync()
 const browser = argv.browser
@@ -28,7 +28,10 @@ const watcher = watch('./src/', { recursive: true }, (event, filename) => {
     root_folder !== undefined && allowed_folders.includes(root_folder)
   if (is_valid_folder) {
     console.log(`Detected ${event} in ${filename}`)
-    build()
+    build({
+      browser,
+      outDir: OUT_DIR,
+    })
   }
 })
 
@@ -44,67 +47,24 @@ console.log(
   '\n-----------------------------------------------------\n'.yellow.bold
 )
 
-const OUT_DIR = 'out'
-
 start()
 
 function start() {
+  if (!is_valid_browser) {
+    process.exit(1)
+    return
+  }
+
   fs.rm(`./${OUT_DIR}`, { force: true })
     .then(() => {
       console.log(`porumai ... ${OUT_DIR} deleted `.yellow)
-      return build()
+      return build({
+        browser,
+        outDir: OUT_DIR,
+      })
     })
     .then(outputs => {
       run_local_extension()
-    })
-}
-
-function build() {
-  const config = get_config(browser as string)
-
-  return Bun.build({
-    entrypoints: ['./src/app/remindoro.tsx', './src/background/index.js'],
-    outdir: `./${OUT_DIR}`,
-    splitting: true,
-    // naming: '[dir]/[name].[ext]',
-    naming: {
-      // default values
-      entry: '[dir]/[name].[ext]',
-      chunk: '[name].[ext]',
-      asset: '[name].[ext]',
-    },
-    define: {
-      'process.env.BUN_RATE_URL': config.rate_url,
-    },
-  })
-    .then(result => {
-      console.log(browser)
-      // const chunks: Array<string> = []
-      const styles: Array<string> = []
-      result.outputs.forEach(value => {
-        // const is_chunk = value.loader === 'js' && value.kind === 'chunk'
-        const is_css =
-          value.loader === 'file' &&
-          value.kind === 'asset' &&
-          value.type.includes('text/css')
-        const file_name = last(value.path.split('/'))
-        if (file_name) {
-          // if (is_chunk) {
-          //   chunks.push(file_name)
-          // } else
-          if (is_css) {
-            styles.push(file_name)
-          }
-        }
-      })
-      copy_image_files()
-      write_html_file(styles)
-      write_manifest_file(browser)
-      console.log(chalk.cyan('porumai ... bun build finished ?'))
-      return new Promise(resolve => resolve(result.outputs))
-    })
-    .catch(error => {
-      console.error('porumai ... bun build error'.red)
     })
 }
 
@@ -129,48 +89,4 @@ function run_local_extension() {
         console.error(err)
       })
   }
-}
-
-function write_html_file(styles: Array<string>) {
-  console.log('porumai ... writing popup.html'.blue)
-  // const chunks_markup = chunks.map(
-  //   chunk => `<script src="./${chunk}" type="module"></script>`
-  // )
-  const styles_markup = styles.map(
-    style => `<link rel="stylesheet" href="./${style}">`
-  )
-
-  const template = `<!DOCTYPE html>
-<head>
-  <meta name="viewport" content="minimum-scale=1, initial-scale=1, width=device-width" />
-  <meta charset="UTF-8" />
-  ${styles_markup}
-</head>
-
-<body>
-  <!-- Porumai: This is our Popup/App shell -->
-  <div id="remindoro-app"></div>
-  <script type="module" src="./app/remindoro.js"></script>
-</body>`
-
-  Bun.write(`./${OUT_DIR}/popup.html`, template)
-}
-
-function write_manifest_file(browser) {
-  console.log(`porumai ... writing manifest for ${browser}`.blue)
-  fs.readFile(`./src/manifests/${browser}/manifest.json`, 'utf8')
-    .then(data => {
-      const manifest = JSON.parse(stripJsonComments(data))
-      manifest.version = process.env.npm_package_version
-      Bun.write(`./${OUT_DIR}/manifest.json`, JSON.stringify(manifest))
-    })
-    .catch(err => {
-      console.error(err)
-    })
-}
-
-function copy_image_files() {
-  fs.cp('./src/img', `./${OUT_DIR}/img`, { recursive: true }).then(() => {
-    console.log('porumai ... img file copied')
-  })
 }
